requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
feedparser==6.0.11
jinja2==3.1.3
python-multipart==0.0.9
aiosqlite==0.19.0
apscheduler==3.10.4
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
python-dotenv==1.0.1 